project structure updated and classes added
