+++ app1.cpp
@@ -1,0 +1,10 @@
+#include <iostream>
+using namespace std;
+
+int main(int args, char *argv[]) // g++ -Wall -o p app1.cpp
+{
+    cout << "args= " << args << " and argv[] " << argv[1] << "    |||;" << endl
+         << endl;
+
+    return 0;
+}
